define macros for byte and us conversion

+++ src/statistics/client_aggregated_statistics.c
@@ -58,7 +58,7 @@ namespace covered
         {
             uint32_t tmp_latency_cnt = client_raw_statistics_ptr->latency_histogram_[latency_us];
             cur_latency_cnt += tmp_latency_cnt;
-            double cur_ratio = 0.0;
+            double cur_ratio = double(0.0);
             if (total_latency_cnt != 0)
             {
                 cur_ratio = static_cast<double>(cur_latency_cnt) / static_cast<double>(total_latency_cnt);
